coordis sql 모델 생성 (INSERT,SELECT)
